config.py
import os
from dotenv import load_dotenv

load_dotenv()

DATABASE_URL        = os.getenv("DATABASE_URL")
GOOGLE_CLIENT_ID    = os.getenv("GOOGLE_CLIENT_ID")
GOOGLE_CLIENT_SECRET= os.getenv("GOOGLE_CLIENT_SECRET")
SECRET_KEY          = os.getenv("SECRET_KEY", "fallback-dev-secret")
